don't allow last admin change/deleted
